requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
python-dotenv
email-validator
pydantic
cryptography